Pagegroup issue fix added
In case when pageGroup pattern is missing, our script was matching topmost pageGorup in the list due to wild card regex nature, we fixed that in this issue. Also added a check before forcing variation that if creationProcess tarted or not, as forcing a variation was causing double trigger for ads.