[FIX] - Auction Feedback missing Prebid Winner details
The issue here was that we get the prebid winner details using _apPbJs.getAdserverTargeting() which might return details of a bid that was received in the previous auction. The bid from previous auction that doesn't win is compared with the current auction bids which is why getAdserverTargeting sometimes give adId from a bid that was received in the previous auction.

When we get adId from getAdserverTargeting, we try to get its details from the bids that were received in the CURRENT AUCTION, which was the issue since previous bid won't be available in such case.